requirements.txt
# Core Flask dependencies
flask>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# Code Q&A dependencies
lancedb>=0.4.0
openai>=1.0.0
tree-sitter>=0.21.3
pandas>=2.0.0
markdown>=3.4.0

# GDD RAG dependencies
dashscope>=1.25.0
lightrag>=0.1.0b6
lightrag-hku>=1.4.9.8
docling>=1.0.0  # PDF to Markdown conversion

# Supabase (for storage)
supabase>=2.0.0
pgvector>=0.2.0

# Optional but recommended
redis>=5.0.0
cohere>=4.0.0

